refactor: separate file and console loggers, test utils
