+++ packages/scripts/src/commands/app-data/convert/utils/index.ts
@@ -1,6 +1,8 @@
-export * from "./app-data-string.utils";
-export * from "./app-data-condition.utils";
 export * from "./app-data-action.utils";
+export * from "./app-data-condition.utils";
+export * from "./app-data-override.utils";
+export * from "./app-data-string.utils";
+
 export * from "./logging";
 // Re-export common shared utils for easier import
 export * from "../../../../utils";
